rename tf-module-defaults to tf-module
